+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - New Command `misc:register:command` to Register Command to service.yaml files.
   - New Command `db:info` Database Table Size (Issue #37).
   - New Command `config:link:environment` links the environment configration files. Ideal for CI/CD.
+  - New Command `config:update:yaml` update the module configuration yaml with the data from the database.
   - ./bin/generate_service_yaml.php to generate `services.yaml` for `oe-console`
   - ./bin/oxrun-light.php small application with one command `misc:register:command`
   - Command `misc:generate:yaml:config` has new Option `--list` to show all configrations